backend/ingestion/requirements.txt
flask==3.1.3
supabase==2.28.0
python-dotenv==1.2.2
openai==2.25.0
pika==1.3.2
flask-cors==5.0.1
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.30.0
structlog>=24.1.0